refactor(scripts): extract common write_outputs and parse_repositories
- Add scripts/common.py with shared write_outputs and parse_repositories
- Refactor build_repos_matrix.py and build_target_operations.py to use common
- Remove ~70 lines of duplicated code across both scripts
- All 31 tests pass

Made-with: Cursor
